Enforce key as non-complex.
If an object’s key is a collection, it’s a big issue. However, most
languages will recuperate from a key that is a primitive or a number
(even though a String is required)… hence, we enforce keys to be
non-complex.

We also protect against “dangling” keys bay adding a NULL value.