Prevent worker initialization from closing a pipe opened by a concurrent thread (in custom `fork`).
The previous logic called `close` twice. Normally this is harmless.
But, if a concurrent thread opens a new `fd` between these two calls,
that `fd` will be closed by the second call.

This was rare enough to ignore (especially as this code runs before any
threads are created), but why ignore if this can be fixed?