Fix kqueue auto-close to prevent `close` being called for valid `fd`
It seems that when `fork` is called, the `kqueue` file descriptor is
automatically closed by the OS.

This means that opening a new file descriptor right after `fork` and
before a call to `evio_close` could result in the new (valid and
desired) file descriptor being unintentionally closed.

This extra test resolves this possible issue (that was driving me nuts
for two weeks now).