Negative `threads` / `processes` settings should have meaningful values
The mathematical logic is obscure, but in essence:

When values are negative, the represent the fraction of the cores to be
used (i.e., -t -2 == cores/2).

However, if both values (threads and processes) are negative (or zero)
and unequal, the numbers are assumed to represent a ratio where 0 is
replaced by -1. An attempt to divide the cores according to the ratio
will be made (i.e., `-t -2`, similar to `-t -4 -w -2`, will result in
ta ratio of 2:1 , having 2 threads per worker, similar to `-t 2 -w -2`).