improve random speed
fio_rand64 isn't meant to be cryptographic.

The previous implementation leveraged SHA-512 and was exteremly slow.

The current implementation uses RiskyHash for seeding (entropy creation) and uses deterministic bit shuffling and automatic re-seeding to provide a significatnly faster random generator.

Tests using the code from http://xoshiro.di.unimi.it/hwd.php show that the randomness is fairly strong for the first 0.5Tb of data or so - probably stronger than required for a non cryptographic PRG.

The approach was modeled after `xoroshiro128+`, though the algorithm is different.